requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python
pydantic
redis
sse-starlette
pydantic-settings
gunicorn
